feat: added text cell in UIComponents
