+++ lm-feedUI-iOS/lm-feedUI-iOS/Source/Classes/UI/Base/Cells/LMFeedBaseDocumentCell.swift
@@ -8,6 +8,13 @@
 import UIKit
 
 open class LMFeedBaseDocumentCell: LMPostWidgetTableViewCell {
+    
+    open private(set) lazy var postText: LMFeedPostBaseTextCell = {
+        let postText = LMUIComponents.shared.textCell.init()
+        postText.translatesAutoresizingMaskIntoConstraints = false
+        return postText
+    }()
+    
     // MARK: UI Elements
     open private(set) lazy var documentContainerStack: LMStackView = {
         let stack = LMStackView().translatesAutoresizingMaskIntoConstraints()
@@ -60,7 +67,7 @@ open class LMFeedBaseDocumentCell: LMPostWidgetTableViewCell {
         postID = data.postID
         userUUID = data.userUUID
                 
-        setupPostText(text: data.postText, showMore: data.isShowMore)
+        postText.configure(text: data.postText, showMore: data.isShowMore)
         
         topicFeed.configure(with: data.topics)
